refactor calendar components: extract utility functions and constants for better modularity
