+++ frontend/packages/app/src/pages/project-details/tabs/calendar/constants.ts
@@ -1,0 +1,9 @@
+// Gantt layout constants
+export const COLUMN_WIDTH = 38; // px per day column
+export const WEEK_LABEL_HEIGHT = 30; // px – week range row
+export const DAY_HEADER_HEIGHT = 30; // px – day numbers row
+export const ROW_HEIGHT = 60; // px – each item row
+export const MIN_CARD_DAYS = 1; // minimum card width in days
+
+// Calendar grid constants
+export const DAY_HEADERS = ["Mon", "Tue", "Wed", "Thu", "Fri", "Sat", "Sun"];
